Split out appearance settings from general section
